Fix ordering issues in requesting backfill
Using a map means the orders are processed in an indeterminate order,
which causes a lot of `/state_ids` requests as the `prev_events` are
not known when the order is lost.

Also topologically sort results from multiple servers when doing a
`RequestBackfill` just to be safe, since otherwise results retrieved
from multiple servers could end up out-of-order too.